+++ lib/Ansifmt.ml
@@ -7,5 +7,4 @@ type color = Color.t
 type styling = Styling.t
 type stylizer = Formatting.Stylizer.t
 
-let make_styling = Styling.create
 let format = Formatting.Util.format